Add build tag to omit NewClient
